Split background into modules
