+++ v2/src/libs/background/index.ts
@@ -1,0 +1,3 @@
+export type { Message, Callback } from './types';
+export { default as authenticate } from './authenticate';
+export { default as signout } from './signout';
