+++ v2/src/libs/background/types.ts
@@ -1,0 +1,9 @@
+/* eslint no-unused-vars: ["error", { "args": "none" }] */
+export type Callback = (args: string | URL | Record<string, string>) => void;
+
+export type Message = {
+  action: 'authenticate' | 'signout',
+  storageKey: string,
+} & {
+  action: string;
+};
